Create e2e/test_form.py
